+++ pickup_radius/Settings.cs
@@ -1,0 +1,28 @@
+﻿using BepInEx.Configuration;
+
+public class Settings {
+    private static Settings m_instance = null;
+    public static Settings Instance {
+        get {
+            if (m_instance == null) {
+                m_instance = new Settings();
+            }
+            return m_instance;
+        }
+    }
+    private DDPlugin m_plugin = null;
+
+    // General
+    public static ConfigEntry<bool> m_enabled;
+    public static ConfigEntry<string> m_log_level;
+    public static ConfigEntry<float> m_pickup_radius;
+
+    public void load(DDPlugin plugin) {
+        this.m_plugin = plugin;
+
+        // General
+        m_enabled = this.m_plugin.Config.Bind<bool>("General", "Enabled", true, "Set to false to disable this mod.");
+        m_log_level = this.m_plugin.Config.Bind<string>("General", "Log Level", "info", "[Advanced] Logging level, one of: 'none' (no logging), 'error' (only errors), 'warn' (errors and warnings), 'info' (normal logging), 'debug' (extra log messages for debugging issues).  Not case sensitive [string, default info].  Debug level not recommended unless you're noticing issues with the mod.  Changes to this setting require an application restart.");
+        m_pickup_radius = this.m_plugin.Config.Bind<float>("General", "Pickup Radius", 2.5f, "Radius (in meters/yards?) around player in which objects will be drawn to the player (float, default 2.5 [game default]).");
+    }
+}