+++ testing/TestingPlugin.cs
@@ -11,6 +11,8 @@ using UnityEngine;
 using UnityEngine.Events;
 using UnityEngine.SceneManagement;
 using Cysharp.Threading.Tasks;
+using UnityEngine.InputSystem;
+using Unity.Mathematics;
 
 public static class PluginInfo {
 
@@ -53,7 +55,139 @@ public class TestingPlugin : DDPlugin {
         }
     }
 
-	/*
+    //TileOccupier, Health, Cuttable
+
+    public class Bulldozer : MonoBehaviour {
+        private const float CHECK_FREQUENCY = 0.5f;
+        private LocalPlayerController m_player;
+        
+        private void Awake() {
+            this.m_player = this.gameObject.GetComponent<LocalPlayerController>();
+            this.StartCoroutine(this.run_routine());
+        }
+
+        private void bulldoze() {
+            try {
+                foreach (GameObject obj in Resources.FindObjectsOfTypeAll<GameObject>()) {
+                    //if (obj.transform.parent == null || !obj.name.StartsWith("CommonFarmTree")) {
+                    //    continue;
+                    //}
+                    //UnityUtils.list_ancestors(obj.transform, _debug_log);
+                    if (!obj.activeSelf) {
+                        continue;
+                    }
+                    float distance = Vector3.Distance(this.m_player.transform.position, obj.transform.position);
+                    if (distance >= 10) {
+                        continue;
+                    }
+                    UnityUtils.list_ancestors(obj.transform, _debug_log);
+                    _debug_log($"[{Time.time}] distance: {distance}");
+                }
+                /*
+                int aura_radius = 10;
+                Level level = this.m_player.Level;
+                if (level == null) {
+                    return;
+                }
+                TileOccupier[,] occupiers = (TileOccupier[,]) ReflectionUtils.get_field_value(level, "m_gameObjects");
+                int2 player_tile = this.transform.position.ToTile(level);
+                foreach (int2 check_pos in new RectangleFillIterator(new int2(player_tile.x - aura_radius, player_tile.y - aura_radius), new int2(player_tile.x + aura_radius, player_tile.y + aura_radius))) {
+                    TileOccupier occupier = occupiers[check_pos.x, check_pos.y];
+                    Health health = null;
+                    Cuttable cuttable = null;
+                    if (occupier == null || (health = occupier.gameObject.GetComponent<Health>()) == null || (cuttable = occupier.gameObject.GetComponent<Cuttable>()) == null) {
+                        continue;
+                    }
+                    _debug_log(occupier);
+                    foreach (Component component in occupier.gameObject.GetComponents<Component>()) {
+                        _debug_log($"--> {component.GetType()}");
+                    }
+                    _debug_log($"healthData: {(HealthData) ReflectionUtils.get_field_value(health, "m_healthData")}");
+                    occupier.transform.position += occupier.transform.up * 2;
+                    //health.ChangeHealth(-float.MaxValue, this.m_player);
+                    //cuttable.OnKilled(this.m_player);
+                }
+                */
+            } catch (Exception e) {
+                _error_log("** Bulldozer.bulldoze ERROR - " + e);
+            }
+        }
+
+        private IEnumerator run_routine() {
+            for (;;) {
+                //this.bulldoze();
+                yield return new WaitForSeconds(CHECK_FREQUENCY);
+            }
+        }
+    }
+
+    [HarmonyPatch(typeof(LocalPlayerController), "Initialize")]
+    class HarmonyPatch_LocalPlayerController_Initialize {
+        private static void Postfix(LevelManager __instance) {
+            __instance.gameObject.AddComponent<Bulldozer>();
+        }
+    }
+    
+    /*
+    private static bool m_added_actions = false;
+    private static bool m_trying_to_jump = false;
+    private static Rigidbody m_rigid_body = null;
+
+    [HarmonyPatch(typeof(LocalPawn), "Awake")]
+	class HarmonyPatch_LocalPawn_Awake {
+		private static bool Prefix(LocalPawn __instance) {
+			try {
+				m_rigid_body = __instance.gameObject.GetComponent<Rigidbody>();
+                _debug_log(__instance.name);
+				return true;
+			} catch (Exception e) {
+				logger.LogError("** HarmonyPatch_LocalPawn_Awake.Prefix ERROR - " + e);
+			}
+			return true;
+		}
+	}
+
+    [HarmonyPatch(typeof(LocalPawn), "Update")]
+    class HarmonyPatch_LocalPawn_Update {
+        private static bool Prefix(LocalPawn __instance) {
+            try {
+                if (m_trying_to_jump) {
+                    m_trying_to_jump = false;
+                    m_rigid_body.isKinematic = false;
+                    m_rigid_body.useGravity = true;
+                    m_rigid_body.mass = 1;
+                    _debug_log($"jumping?  mass: {m_rigid_body.mass}, isKinematic: {m_rigid_body.isKinematic}, useGravity: {m_rigid_body.useGravity}");
+                    m_rigid_body.AddForce(m_rigid_body.transform.up * 10, ForceMode.Impulse);
+                }
+                return true;
+            } catch (Exception e) {
+                logger.LogError("** HarmonyPatch_LocalPawn_Update.Prefix ERROR - " + e);
+            }
+            return true;
+        }
+    }
+
+    [HarmonyPatch(typeof(LocalPlayerController), "Initialize")]
+    class HarmonyPatch_LocalPlayerController_Initialize {
+        private static void Postfix(LocalPlayerController __instance, PlayerInput ___m_input) {
+            try {
+                if (m_added_actions) {
+                    return;
+                }
+                m_added_actions = true;
+                UnityUtils.dump_control_paths();
+                UnityUtils.add_hotkey_action(___m_input, "jump", "", "/Keyboard/space", performed: context => {
+                    m_trying_to_jump = true;
+                    _debug_log($"JUMP button pressed");
+                });
+            } catch (Exception e) {
+                logger.LogError("** HarmonyPatch_LocalPlayerController_Initialize.Prefix ERROR - " + e);
+            }
+        }
+    }
+    */
+
+    /*
 	[HarmonyPatch(typeof(), "")]
 	class HarmonyPatch_ {
 		private static bool Prefix() {
@@ -74,7 +208,7 @@ public class TestingPlugin : DDPlugin {
 		private static bool Prefix() {
 			try {
 
-				return false;
+				return true;
 			} catch (Exception e) {
 				logger.LogError("** XXXXX.Prefix ERROR - " + e);
 			}