+++ camera_control/changelog.md
@@ -1,4 +1,16 @@
-﻿# 0.0.3
+﻿# 0.0.6
+## Features
+- Added settings and hotkeys for changing the camera's "look" angle
+
+# 0.0.5
+## Features
+- Added setting (defaults to true) that disables the up/down drift of camera to a specific angle when zooming
+
+# 0.0.4
+## Bug Fixes
+- Fixed a glitch causing blueprint objects to not be visible if camera fog is disabled
+
+# 0.0.3
 ## Features
 - Added setting to remove Unity distance fog, since it really isn't required for performance in this game